First cut at using Vagrant for development
I imported and adapted the Vagrantfile and provisioning shell script
from Agile Ventures Local Support. The provisioning script imports the
generic Rails script from AgileVentures/setup-scripts.

This is held together with duct tape and string, but it works, so it is
ready for others to try out. Then we can make improvements based on
experience.