remove flaky javascript time control
The timepicker doesn't seem to work right... or maybe we weren't using it right.  However, I haven't been able to get it fail with cucumber, despite following the same steps as in Chrome to get it to fail.  i'm going to go with the default rails helper time/date pickers for now

The timepicker doesn't seem to work right... or maybe we weren't using it right.  However, I haven't been able to get it fail with cucumber, despite following the same steps as in Chrome to get it to fail.  i'm going to go with the default rails helper time/date pickers for now

The timepicker doesn't seem to work right... or maybe we weren't using it right.  However, I haven't been able to get it fail with cucumber, despite following the same steps as in Chrome to get it to fail.  i'm going to go with the default rails helper time/date pickers for now

The timepicker doesn't seem to work right... or maybe we weren't using it right.  However, I haven't been able to get it fail with cucumber, despite following the same steps as in Chrome to get it to fail.  i'm going to go with the default rails helper time/date pickers for now

Reverted welcome page/message about scrum times.

modified view spec test to reflect new controls for event date and start time and end time

not working, but hoping to get feedback on the feature stepsof the form:
And(/^I fill in event year with "([^"]*)"$/) do |year|
  select year, :from => "event_event_date_1i"
end

Reverted the date back to use the bootstrap-datepicker
Then reverted the "fill in the event field" step of the test with the date in it, which removed the need to set the date_select fields in Capybara steps.

reverted part of a view spec test for the event_date to once again be a datepicker field

commit to force another travis test

reformatting (undo rubymine's auto-formatting) closer to original
small change to events _form view  (added back variable for event_date)
more descriptive name for feature

fixed timepicker!
Yay!  The problem was simply that the timepicker wanted the time in the same format as it uses, so am/pm was necessary.  I changed back the code that I had previously changed.  Also, I added a small enhancement, that of initializing it with today's time and date for the start and a half hour later for the end time.  This is useful for events.

Moved format_timepicker from class method into helper

forcing travis re-test
Also, simplified text_field by removing data: format: