Retrieve EventInstance once for all users
Moved the call to ActiveRecord EventInstance out of the loop within
KarmaCalculator to prevent hitting the database for each user.

This will relieve the workload the database needs to do in order for us
to calculate a user's karma more efficiently. There have been times
where the server kills the task due to the rake task taking to long.