Add exception handling to KarmaCalculations
Wrapped KarmaCalculator into a rescue clause.

This ensures the karma calculations completes for all users in the case
that there is a standard error. This will also allow us to log errors if
one occurrs while processing a particular user's karma by logging the
user's id along with the error message and backtrace.