Add tests to show event modifier info on events page
Added cucumber tests for the event show page to check that both creator
and modifier info exists on the page after an event is edited.

Currently when someone modifies an event the creator info is destroyed.
This is needed so that we can preserve the creator info and show who
modified the event.