Add feature spec
Added feature tests for when there are zero, 3, and six events to
display on the show page. Also testing for the text inside the message,
and testing the message when there are 0 events to display.

Added feature to check order of events and to check that we're not
displaying event instances that do not have a youtube link.

This feature is being added so that a user can easily tell which events
are active