Configure Warden
Configured warden's built in stubbing actions.

This is done so that we can have stub devise sign in on our acceptance
tests without all the extra overhead of driving a browser.

Multiple tests rely on a user being signed and being able to speed up
this action will significantly increase our run time.