Reword background steps
Using a declaritive syntax for the back ground steps.

This is done because background steps shouldn't be checking
functionality they're main purpose is to declare the state of the system
before the test. The background steps were slowing down each test,
because we were driving capybara to test the background steps on each
test.

Rewording to implement steps that do not drive the browser.