Create new event step
Create a new event with the option to repeat.

This method is to replace background steps that were driving the browser
to create an event before a suite of tests in the edit past event
feature.

Since we don't need to test the functionality of creating an event for
that feature we can use this method to speed up the tests.