Update tests now that source repo is not built
There are three tests failing due to the source repository not being
built for every project regardless if there is a valid github link.

Source repo's should only exist when there are github links associated
to a project.

These three tests were trying to add github links to a project that
didn't previously had a github link, hence the button needed to be
clicked before the value could be filled in.