Update stub to include method
Stubbed method that calls the API instead of only the API call. The
other way of stubbing is fine, but since there are 'puts' statements
inside the method it makes reading the test output confusing and/or
inconvenient.

Updated the Regex to include a plural form.

Updated the JSON file