Refactor connection view
Created method for the slack channel link and used it in the view
instead of hard-coding the link.

This is done so that values that may change are hard-coded in one
location. It will also help to dry out test code since an object can be
asked how to get to its slack channel.