2503 refactor connections feature (#2514)
* Refactor connection spec

Used the scenario outline to test for when projects do not have any
connections.

This helps to dry out the spec since it used the same scenario 3 times
with different keyword values.

* Refactor connection view

Created method for the slack channel link and used it in the view
instead of hard-coding the link.

This is done so that values that may change are hard-coded in one
location. It will also help to dry out test code since an object can be
asked how to get to its slack channel.

* Create a universtal step definition

Created a step definition that can be used to assert all API links a
project is connected to.

* Implement scenario outline for common scenarios

* Update slack steps to use new universal step

Removed the slack step for asserting the slack link and updated the
feature steps to use the new universal step for asserting links.

* Update pivotal tracker steps

Updated the assertions for pivotal tracker links to use the new step
assertion.

* Remove duplicate step definition

Removed the link assertion step and updated the feature spec to use the
'basic_step' link assertion.