Return false unless current start time exists
Returning false unless the events current start time exists.

When an event exists before the actual event start date the event
occurrence method is returning nil.

This is happening when a user creates an event and starts a hangout, but
then after the hangout finishes the event is edited (most likely no one
showed and so they reschedule) to start at a later date.