Update Faraday gem
Removed twitter gem and associated code.

It was mentioned that twitter is no longer being used to post events to,
and that the gem and associated code could be removed completely.

This version of the twitter gem was preventing the Faraday and stripe
gem from being updated.

Updated stripe to latest version so that the faraday gem can be updated
in order to fix the fixnum deprecation warnings.