Update null user to have an id (#2588)
* Update null user to have an id

Updated the null user class to have an id.

This allows removal of the conditional in the view, because the
<link_to> method can now render sincet the returned NullUser object has
an id.

We should create an anonymous/ghost user in production with an id of -1.

Fixes #2583

* Add spec for viewing anonymous show page

This is done as additional verification for viewing the anonymous page
with id of -1.

* Create rake task for adding an anonymous user

This is done to reduce the risk of deployment issues accross staging and production databases. Also added rake task to `seeds.rb` to match
production and development environment.