Update live event test to use recent event
Updated this test to use a recent event instead of an event that started
in 2014. Added factories to implement this change to have a
recent_event, and a live_event_instance. And then updated the test to
use the live_event_instance which associates to the recent_event that
started 8 hours ago.

This is being done to update the test to use a more realistic scenario,
because the current scenario is checking that an event started in 2014
which is still being recorded shows up as an upcoming event.
Realistically an event may be recorded for a few hours at most, and a
recording "probably" would never last more than a day, hence the reason
in the previous commit to pull events from the previous day, however
it's causing a failure here, because this non-recurring event had a
start date from 2014.