Update cuke step to create event w/req fields
Updated the capybara step to create recurring events with repeat_ends
set to false. This is being done because repeat_ends should either be
true or false. It's true when a non-recurring event is creating and when
a recurring event is created with an end date; otherwise it's false.

Added issue to add validation for this field to prevent future issues
regarding this. When events are created through the website this is not
an issue, but if its created programmatically then this field can be
bypassed when it shouldn't.