Append error message to array
The messages hash has a value listing all errors in an array. Updated
Image validator to append error messages to the array instead of over
writing the array with a String. This is being done because it causes
Rails ActiveRecord to error out when it does post validation checks,
because it's expecting an array, but instead receives a String.

Fixes #2857