Append error message to array (#2929)
* Add test to check custom image validator

This test checks our custom image url validation is appending the errors
correctly to the error messages hash. This is needed because some users
see a 500 internal server error when creating a project with an invalid
url.

* Append error message to array

The messages hash has a value listing all errors in an array. Updated
Image validator to append error messages to the array instead of over
writing the array with a String. This is being done because it causes
Rails ActiveRecord to error out when it does post validation checks,
because it's expecting an array, but instead receives a String.

Fixes #2857