Add test for scrum that doesn't end
Added test for an event that never ends, but has an end date.
(ex: repeat = false and repeat_ends_on 'some date in the past')

This is being done to emulate an actual current event in the database
due to the repeat_ends_on date not being updated when an event is
switched from a recurring event with an end date to a recurring event
without an end date.