Decrease time frame on schedule instance
Decreased the returned time frame when a schedule instance is created
for current_start_time method.

This is being done because the current_start_time method is
determining if there is a current event_instance, however if a start
time isn't specified it grabs all event instances associated to the
event, and then checks each one to determine if there is a live event.
If an event instance is live it would be the most recent event so
narrowing the time frame when creating a Schedule instance speeds up the
load time on for the events page.