Return current time from the most recent event
Use occurrences_between method to retrieve the most recent event within
the last month.

The previous_occurrence(from) uses the actual start_date of the event
and pulls all events from the start_date of the event until the time
passed to the methods from parameter, and then returns the last event.
We are passing in Time.current to it and so the method finds all
instances of the event just to return the most recent.