Fix future events query (#2922)
* Create unit tests for Event.future_events

Created unit tests for the future_events method to test a theory that
events in the past are being returned when they shouldn't. These tests
will also help test out future implementations and prevent regression.

* Add one more spec to test new query

This is being added to prevent regression and ensure query string is
retrieving all future events.

* Update event query to pull active events

Updates event query to pull events with a start date greater than
yesterday, recurring events without an end date, and recurring events
with an end date greater than yesterday. This query retrieves all active
events as of the day before. The reason why it pulls in events from
current time minus one day is because an event that occured a few hours
ago may still be in progress. The returned query results are handed off
to the scheduler to determine actual upcoming events based on any
current event instances and upcoming schedule.

This is needed, because the previous query was pulling in inactive
events causing a bottle neck in the ice cube gem that handles all the
scheduling which calculates if there are events in progress and/or when
the next event if any would occur.

* Update live event test to use recent event

Updated this test to use a recent event instead of an event that started
in 2014. Added factories to implement this change to have a
recent_event, and a live_event_instance. And then updated the test to
use the live_event_instance which associates to the recent_event that
started 8 hours ago.

This is being done to update the test to use a more realistic scenario,
because the current scenario is checking that an event started in 2014
which is still being recorded shows up as an upcoming event.
Realistically an event may be recorded for a few hours at most, and a
recording "probably" would never last more than a day, hence the reason
in the previous commit to pull events from the previous day, however
it's causing a failure here, because this non-recurring event had a
start date from 2014.

* Remove fragment caching

This removes the caching implemented in pr #2824. This is being done,
because it should no longer be necessary to cache results with the
updated event query.

* Update cuke step to create event w/req fields

Updated the capybara step to create recurring events with repeat_ends
set to false. This is being done because repeat_ends should either be
true or false. It's true when a non-recurring event is creating and when
a recurring event is created with an end date; otherwise it's false.

Added issue to add validation for this field to prevent future issues
regarding this. When events are created through the website this is not
an issue, but if its created programmatically then this field can be
bypassed when it shouldn't.

* Add test for scrum that doesn't end

Added test for an event that never ends, but has an end date.
(ex: repeat = false and repeat_ends_on 'some date in the past')

This is being done to emulate an actual current event in the database
due to the repeat_ends_on date not being updated when an event is
switched from a recurring event with an end date to a recurring event
without an end date.

* Update query to pass new test

This updated query also grabs any events that is currently repeating,
but has an end date in the past.
(repeat_ends = false AND repeats != 'never')

* Add to for an event that used to repeat

This test is being added to emulate a test that used to repeat without
an end date, but was updated to an event that never repeats. This is
being added, because the current Event query allows this inactive event
through.

* Update query string

This query is updated to not allow events that used to repeat forever
but are no longer a repeating event.

* Decrease time frame on schedule instance

Decreased the returned time frame when a schedule instance is created
for current_start_time method.

This is being done because the current_start_time method is
determining if there is a current event_instance, however if a start
time isn't specified it grabs all event instances associated to the
event, and then checks each one to determine if there is a live event.
If an event instance is live it would be the most recent event so
narrowing the time frame when creating a Schedule instance speeds up the
load time on for the events page.

* Fix outdated/faulty tests

Updated tests to not use delorean. The delorean was set for a different
date than the tests were being tested for.

* Revert schedule start time

Reverted schedule modifications made in the previous commit as this is
not narrowing the time frame for ice cube to search through.
The schedule.rb ice_cube class still grabs the events start time.

* Return current time from the most recent event

Use occurrences_between method to retrieve the most recent event within
the last month.

The previous_occurrence(from) uses the actual start_date of the event
and pulls all events from the start_date of the event until the time
passed to the methods from parameter, and then returns the last event.
We are passing in Time.current to it and so the method finds all
instances of the event just to return the most recent.