Delete spec that always fails on Sunday
Deleted the spec that is testing the less_than_ten_till_start method.
This method is already under unit test for both positive and negative
paths. This spec fails on Sunday because the event under test has a
Sunday recurring day and current time, however the ice_cube gem returns
a current schedule regardless if the event has ended. This doesn't
happen with system tests and in production as only the events that are
active runs through this method.