send users welcome emails when they join projects (#2646)
* initial feature draft

* work in progress

* Add fixture and tests mailing project joinee

* Add unique message for website one project

* Update feature spec

Updated feature spec to handle multiple emails being sent at the same
time.

These changes are done as two emails are now sent. One to the creator
and one to the joinee.

* Send email if joinee allows notifications

* Update fixtures

Fixtures added and updated for testing the body of the messages.

* Add/Update rspec tests for joinee notifications

Added two seperate tests. One for testing when a user follows a
websiteone project, and another test for the other projects.

* Add/Update mailer views

* Generalize feature spec

* Set host URL

Set the host URL so that links can be used in the mailer as documented
in the Rails guide:
https://guides.rubyonrails.org/action_mailer_basics.html#generating-urls-in-action-mailer-views

* Generate project mailer

* Refactor to seperate project mailer

Moved project mailer logic to its own controller.

* Minor spec updates

* Setup preview for mailers

Mail messages can be viewed at /rails/mailers/project_mailer

* Remove model spec testing mailer actions

This is under test within the system tests.

* Update sandbox email intereceptor

* Delete spec that always fails on Sunday

Deleted the spec that is testing the less_than_ten_till_start method.
This method is already under unit test for both positive and negative
paths. This spec fails on Sunday because the event under test has a
Sunday recurring day and current time, however the ice_cube gem returns
a current schedule regardless if the event has ended. This doesn't
happen with system tests and in production as only the events that are
active runs through this method.