Remove obsolete tests
These tests are assuming the hangout is pinging the app to indicate the
event is still live. However the hangout is no longer connected to the
app. To my knowledge this was done a couple years ago due to changes in
the youtube/google service that was being used.

This explains why the update link allowed anyone to create/update event
instances which is the root cause of the error's being seen.